Add reusable Price component
